requirements.txt
streamlit==1.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.25
pandas==2.1.4
